(refactor): seperated validators into a different file from the useInput hook
